requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.62.0
basicsr>=1.4.2
facexlib>=0.3.0
gfpgan>=1.3.0
realesrgan>=0.3.0
cupy-cuda12x; sys_platform != 'darwin'